+++ src/clojure_lsp/feature/completion_snippet.clj
@@ -1,4 +1,10 @@
-(ns clojure-lsp.feature.completion-snippet)
+(ns clojure-lsp.feature.completion-snippet
+  (:require
+    [clojure-lsp.shared :as shared]
+    [clojure.string :as string]
+    [rewrite-clj.zip :as z]
+    [taoensso.timbre :as log]
+    [clojure-lsp.parser :as parser]))
 
 (defn known-snippets [settings]
   [{:label "comment$"
@@ -61,3 +67,26 @@
    {:label "use$"
     :detail "Create use"
     :insert-text "(:use [${1:namespace} :only [$0]])"}])
+
+(defn replace-snippets-vars [snippet next-loc]
+  (let [current-sexpr (or (some-> next-loc z/string)
+                          "")]
+    (string/replace snippet "$current-form" current-sexpr)))
+
+(defn build-additional-snippets [cursor-loc next-loc settings]
+  (->> (get settings :additional-snippets [])
+       (filter #(or (not (string/includes? (:snippet %) "$current-form"))
+                    (and cursor-loc
+                         next-loc)))
+       (map (fn [{:keys [name detail snippet]}]
+              (if (string/includes? snippet "$current-form")
+                (let [range (shared/->range (meta (z/node next-loc)))]
+                  {:label name
+                   :detail detail
+                   :text-edit {:range (if (= :token (z/tag cursor-loc))
+                                        (update-in range [:start :character] - (count (z/string cursor-loc)))
+                                        range)
+                               :new-text (replace-snippets-vars snippet next-loc)}})
+                {:label name
+                 :detail detail
+                 :insert-text (replace-snippets-vars snippet next-loc)})))))